Add styling for the Header and landing page, Add landingPage folder and add error and its component, Add UsersList compoonent
