+++ app/components/SignoutModal.jsx
@@ -6,11 +6,11 @@ const SignoutModal = ({ isOpen, title, description, confirmText, onCancel, onCon
   return (
     <div className="fixed inset-0 flex items-center justify-center bg-gray-500 bg-opacity-75"> 
       <div className="bg-white p-6 rounded-md w-1/2">
-        <h2 className="text-2xl font-bold mb-4">{title}</h2>
+        <h2 className="text-2xl font-bold mb-4 text-red-500">{title}</h2>
         <p className="text-gray-700 mb-4">{description}</p>
         <div className="flex justify-end gap-2">
           <button
-            className="bg-primary text-white px-4 py-2 rounded-md"
+            className="bg-red-700 text-white px-4 py-2 rounded-md"
             onClick={onConfirm}
           >
             {confirmText}
