+++ app/components/landingPage/LandingPage.jsx
@@ -15,7 +15,7 @@ const LandingPage = () => {
     <div className="-mt-4 pt-10 p-5 flex items-center justify-center">
       <div className="text-center">
         <h1 className="text-4xl font-bold text-primary mb-4">
-          Discover Your Memories with Our Photo Album App
+          Discover Your Memories With Our Photo Album App
         </h1>
         <p className="text-gray-700 text-md md:text-lg mb-2">
           Explore and organize your memories with our photo album application.
@@ -27,7 +27,7 @@ const LandingPage = () => {
         <div className="flex justify-center">
           <button
             onClick={handleGithubSignIn}
-            className="flex items-center justify-center gap-2 bg-gray-400 rounded-lg w-2/3 text-white px-4 py-2 font-semibold hover:bg-opacity-80"
+            className="flex items-center justify-center gap-2 bg-primary rounded-lg w-2/3 text-white px-4 py-2 font-semibold hover:bg-opacity-80"
           >
             <FaGithub className="text-lg"/>
             Continue with Github
