--- app/login/page.jsx
@@ -1,7 +1,0 @@
-const LoginPage = () => {
-  return (
-    <div>LoginPage</div>
-  )
-}
-
-export default LoginPage
